Move specimen import to its own pipeline provider
