+++ study/src/org/labkey/study/pipeline/StudyPipeline.java
@@ -22,16 +22,12 @@ import org.labkey.api.pipeline.PipelineAction;
 import org.labkey.api.pipeline.PipelineDirectory;
 import org.labkey.api.pipeline.PipelineProvider;
 import org.labkey.api.security.permissions.InsertPermission;
-import org.labkey.api.specimen.pipeline.SpecimenBatch;
-import org.labkey.api.study.SpecimenService;
-import org.labkey.api.study.SpecimenTransform;
 import org.labkey.api.study.Study;
 import org.labkey.api.util.FileUtil;
 import org.labkey.api.util.Path;
 import org.labkey.api.view.ActionURL;
 import org.labkey.api.view.ViewContext;
 import org.labkey.study.controllers.StudyController;
-import org.labkey.study.controllers.specimen.SpecimenController;
 import org.labkey.study.model.StudyManager;
 
 import java.io.File;
@@ -76,28 +72,6 @@ public class StudyPipeline extends PipelineProvider
         });
 
         handleDatasetFiles(context, study, directory, files, includeAll);
-
-        files = directory.listFiles(new FileEntryFilter()
-        {
-            @Override
-            public boolean accept(File f)
-            {
-                if (SpecimenBatch.ARCHIVE_FILE_TYPE.isType(f))
-                    return true;
-                else
-                {
-                    for (SpecimenTransform transform : SpecimenService.get().getSpecimenTransforms(context.getContainer()))
-                    {
-                        if (transform.getFileType().isType(f))
-                            return true;
-                    }
-                }
-                return false;
-            }
-        });
-
-        String actionId = createActionId(SpecimenController.ImportSpecimenData.class, "Import Specimen Data");
-        addAction(actionId, SpecimenController.ImportSpecimenData.class, "Import Specimen Data", directory, files, true, false, includeAll);
     }
